+++ src/controllers/index.js
@@ -1,2 +1,3 @@
-export {default as MapController} from './map-controller/map-controller';
-export {default as OrbitController} from './orbit-controller/orbit-controller';
+// Deprecated exports, just to avoid breaking the 4.1 examples
+export {default as MapController} from '../react/controllers/map-controller';
+export {default as OrbitController} from '../react/controllers/orbit-controller';